Parser: simplified out-of-order insertion algorithm
array_splice() works by copying the whole array, element by element, regardless of the index for the insertion. Tags are usually added at the top or inserted near the top of the stack, so copying every element is relatively inefficient for our use-case. This simplified algorithm copies tags manually starting from the top and stops at the index of the insertion. This is much worse than the native array_splice() if the index was in the bottom half of the stack but in most cases this turns out to be at least as performant and scale better with big stacks. (>1000 tags)